Refactored MulticastConfig to be a Java bean with getters and setters.
git-svn-id: http://hazelcast.googlecode.com/svn/trunk@316 3f8e66b6-ca9d-11dd-a2b5-e5f827957e07

+++ src/com/hazelcast/config/ExecutorConfig.java
@@ -7,7 +7,7 @@ public final class ExecutorConfig {
 
 	public final static int DEFAULT_CORE_POOL_SIZE = 10;
 	public final static int DEFAULT_MAX_POOL_SIZE = 50;
-	public final static int DEFAULT_KEEPALIVE_SECONDS = 60;
+	public final static long DEFAULT_KEEPALIVE_SECONDS = 60;
 	
     private int corePoolSize = DEFAULT_CORE_POOL_SIZE;
